requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
itsdangerous==2.1.2
openpyxl==3.1.2
python-pptx==0.6.21
python-docx==0.8.11
